fix: raise catchable OverflowError on integer overflow (council B1/B2/M1/M2)
Address the round-2 council review of PR #30. The P0-3 overflow work relied on
`overflow-checks`, whose panic PyO3 surfaces as an uncatchable `PanicException`
(a `BaseException`); it also missed the JIT and importable-crate paths.

B1/M2 — checked arithmetic everywhere:
- Generated integer `+`/`-`/`*` now lowers to `__rextio_checked_{add,sub,mul}`
  helpers that raise `OverflowError` (pyo3) / return `RextioError` (crate) on
  overflow instead of wrapping or panicking. One shared, mode-aware path fixes
  both the extension module (B1) and the Rust-importable crate (M2), so the
  guarantee travels with the code even when the crate is consumed as a
  dependency (where `[profile.release]` is ignored).
- Bind for-loop variable types during rendering so loop accumulation
  (`acc += x`) is checked too, matching the comprehension path.
- Release builds keep `overflow-checks = true` as a backstop.

M1 — JIT overflow:
- Refuse JIT for int helpers with overflow-prone arithmetic (the Cranelift path
  emits wrapping ops and cannot raise OverflowError); they stay on the checked
  native path. Float scalar helpers remain JIT-eligible; JIT tests migrated to
  the float path and a regression test asserts int arithmetic is not JIT'd.

B2 — e2e marker discipline / CI scope:
- The `e2e` CI job now runs `pytest tests/e2e` instead of `-m needs_cargo`, which
  silently dropped toolchain-free e2e tests (CLI smoke, zipapp) so they ran in no
  CI lane. The conftest auto-skips genuinely-absent toolchains.
- conftest classifies by `_`-delimited segments (not loose substring) and fails
  collection if an e2e test is neither toolchain-tagged nor toolchain-free.

Minors:
- Pin the overflow e2e to `OverflowError` (was `BaseException`) and add loop and
  catchable-as-Exception cases.
- `.gitattributes` pins golden snapshots to LF and the snapshot compare
  normalizes newlines (CRLF safety).
- CI guards against residual `rextio_runtime` references.

Verified: 227 unit/integration passed (81% cov), ruff + mypy clean, full e2e
19 passed / 1 skipped (nuitka), including a real-cargo build raising a catchable
OverflowError on single-multiply and loop accumulation.

+++ tests/e2e/test_native_jit_real_cargo.py
@@ -37,12 +37,12 @@ hot_threshold = 1
         """
 import rextio
 
-def helper(x: int) -> int:
-    return x * 2
+def helper(x: float) -> float:
+    return x * 2.0
 
 @rextio.native
-def compute(x: int) -> int:
-    return helper(x) + 1
+def compute(x: float) -> float:
+    return helper(x) + 1.0
 """,
         encoding="utf-8",
     )
@@ -71,7 +71,7 @@ def compute(x: int) -> int:
         sys.modules.pop(module_name, None)
 
     native_module = importlib.import_module("_rextio_native")
-    assert native_module.jit_app__math_ops__compute(3) == 7
+    assert native_module.jit_app__math_ops__compute(3.0) == 7.0
 
     module = importlib.import_module("jit_app.math_ops")
-    assert module.compute(4) == 9
+    assert module.compute(4.0) == 9.0
